+++ Assets/MyXonotic/Editor/LocalTests.cs
@@ -30,6 +30,7 @@ namespace MyXonotic.EditorTools
             Check(ArenaMath.SplashDamage(5, 5, 90) == 0, "splash radius cutoff");
             Dev14PhysicsTests.Run(Check);
             Dev15WeaponTests.Run(Check);
+            Dev16BotTests.Run(Check);
             LocalBuild.CreateDevelopmentScene();
             var bootstrap = UnityEngine.Object.FindObjectOfType<ArenaBootstrap>();
             Check(bootstrap != null, "serialized bootstrap");